# backend/openapi.yaml
openapi: 3.0.0
info:
  title: PropertyPilot Backend API
  description: Real estate investment analysis platform with financial calculations, property management, and reporting.
  version: 1.0.0
  contact:
    name: PropertyPilot Support
    email: support@propertypilot.com
  license:
    name: MIT
servers:
  - url: https://api.propertypilot.com
    description: Production server
  - url: http://localhost:3001
    description: Development server
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user with email and password, returns session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: sessionId=abc123; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Destroy the current session
      security:
        - SessionAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Logged out successfully
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Returns the authenticated user profile
      security:
        - SessionAuth: []
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/password:
    put:
      tags:
        - Authentication
      summary: Change password
      description: Update the current user's password
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  minLength: 1
                newPassword:
                  type: string
                  minLength: 8
              required:
                - currentPassword
                - newPassword
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Password changed successfully
...
  /api/dashboard/kpi:
    get:
      tags:
        - Dashboard
      summary: Get portfolio KPIs
      description: Get key performance indicators for the user's portfolio
      security:
        - SessionAuth: []
      responses:
        '200':
          description: KPI data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      kpi:
                        $ref: '#/components/schemas/PortfolioKPI'
  /api/dashboard/portfolio-trend:
    get:
      tags:
        - Dashboard
      summary: Get portfolio trend
      description: Get historical portfolio trends
      security:
        - SessionAuth: []
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            default: 12
      responses:
        '200':
          description: Trend data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      trend:
                        type: array
                        items:
                          $ref: '#/components/schemas/TrendPoint'
  /api/dashboard/deal-flow:
    get:
      tags:
        - Dashboard
      summary: Get deal flow
      description: Get deal pipeline statistics
      security:
        - SessionAuth: []
      responses:
        '200':
          description: Deal flow data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      dealFlow:
                        $ref: '#/components/schemas/DealFlow'
...
  /api/integrations:
    get:
      tags:
        - Integrations
      summary: List integrations
      description: Get all available integrations and their status
      security:
        - SessionAuth: []
      responses:
        '200':
          description: List of integrations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      integrations:
                        type: array
                        items:
                          $ref: '#/components/schemas/Integration'
  /api/integrations/{service}/connect:
    post:
      tags:
        - Integrations
      summary: Connect integration
      description: Connect to an external service
      security:
        - SessionAuth: []
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
            enum: [zillow, airdna, mls]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                apiKey:
                  type: string
                apiSecret:
                  type: string
              required:
                - apiKey
      responses:
        '201':
          description: Integration connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      integration:
                        $ref: '#/components/schemas/Integration'
  /api/integrations/{service}/disconnect:
    delete:
      tags:
        - Integrations
      summary: Disconnect integration
      description: Disconnect from an external service
      security:
        - SessionAuth: []
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
            enum: [zillow, airdna, mls]
      responses:
        '200':
          description: Integration disconnected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Integration disconnected successfully
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session-based authentication via httpOnly cookie
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
      required:
        - email
        - password
        - name
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
...
    PortfolioKPI:
      type: object
      properties:
        totalProperties:
          type: integer
        totalInvested:
          type: number
          format: float
        totalValue:
          type: number
          format: float
        averageCapRate:
          type: number
          format: float
        totalMonthlyNOI:
          type: number
          format: float
        underContract:
          type: integer
        activeDeals:
          type: integer
    TrendPoint:
      type: object
      properties:
        month:
          type: string
          format: date
        totalValue:
          type: number
          format: float
        monthlyNOI:
          type: number
          format: float
        propertyCount:
          type: integer
    DealFlow:
      type: object
      properties:
        lead:
          type: integer
        analyzing:
          type: integer
        offer:
          type: integer
        under_contract:
          type: integer
        owned:
          type: integer
        sold:
          type: integer
...
    Integration:
      type: object
      properties:
        id:
          type: string
          format: uuid
        service:
          type: string
        isActive:
          type: boolean
        lastSyncAt:
          type: string
          format: date-time
          nullable: true
        syncStatus:
          type: string
          nullable: true
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Validation failed
              details:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: array
                      items:
                        type: string
                    message:
                      type: string
              requestId:
                type: string
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Unauthorized
              requestId:
                type: string
...